rides service feign requestInterceptor fix
