--- rides-service/src/main/java/io/simakkoi9/ridesservice/config/FeignConfig.java
@@ -1,49 +1,0 @@
-package io.simakkoi9.ridesservice.config;
-
-import feign.Capability;
-import feign.Logger;
-import feign.RequestInterceptor;
-import feign.Retryer;
-import feign.micrometer.MicrometerCapability;
-import io.micrometer.core.instrument.MeterRegistry;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.context.request.RequestContextHolder;
-import org.springframework.web.context.request.ServletRequestAttributes;
-
-@Configuration
-public class FeignConfig {
-
-    @Bean
-    public Retryer retryer() {
-        return Retryer.NEVER_RETRY;
-    }
-
-    @Bean
-    public Logger.Level feignLoggerLevel() {
-        return Logger.Level.FULL;
-    }
-
-    @Bean
-    public Capability capability(final MeterRegistry registry) {
-        return new MicrometerCapability(registry);
-    }
-
-    @Bean
-    public RequestInterceptor requestInterceptor() {
-        return requestTemplate -> {
-            ServletRequestAttributes attributes =
-                    (ServletRequestAttributes) RequestContextHolder.getRequestAttributes();
-            if (attributes != null) {
-                String userId = attributes.getRequest().getHeader("X-User-Id");
-                String userRole = attributes.getRequest().getHeader("X-User-Role");
-                if (userId != null) {
-                    requestTemplate.header("X-User-Id", userId);
-                }
-                if (userRole != null) {
-                    requestTemplate.header("X-User-Role", userRole);
-                }
-            }
-        };
-    }
-}
