+++ rides-service/src/main/java/io/simakkoi9/ridesservice/exception/PassengerAccessDeniedException.java
@@ -1,0 +1,9 @@
+package io.simakkoi9.ridesservice.exception;
+
+import org.springframework.context.MessageSource;
+
+public class PassengerAccessDeniedException extends CustomRuntimeException {
+    public PassengerAccessDeniedException(String messageKey, MessageSource messageSource, Object... args) {
+        super(messageKey, messageSource, args);
+    }
+}
