+++ rides-service/src/main/resources/messages_en.properties
@@ -7,6 +7,7 @@ page.current.min=Invalid current page value.
 page.size.min=Invalid page size value.
 distance-processing.error=Error calculating distance. Please try again later.
 ride.not-found.error=Requested ride with ID {0} was not found.
+passenger.access.denied.error=Access denied for passenger.
 passenger.not-found.error=Requested passenger not found
 passenger.not-available.error=Passenger service connection failed.
 passenger.temp.not-available.error=Passenger service temporary not available.
