+++ rides-service/src/main/resources/messages_ru.properties
@@ -6,6 +6,7 @@ address.invalid=Указан некорректный адрес. Проверь
 page.current.min=Неверный номер страницы.
 page.size.min=Неверный размер страницы.
 distance-processing.error=Ошибка при расчёте расстояния. Попробуйте позже.
+passenger.access.denied.error=Доступ для данного пассажира воспрещен.
 ride.not-found.error=Запрошенная поездка по ID {0} не найдена.
 passenger.not-found.error=Запрошенный пассажир не найден.
 passenger.not-available.error=Ошибка подключения к сервису пассажиров.
